Rename the `classylimelights` package to `limelight`
