+++ src/main/java/io/github/roboblazers7617/limelight/PoseEstimate.java
@@ -1,8 +1,8 @@
-package io.github.roboblazers7617.classylimelights;
+package io.github.roboblazers7617.limelight;
 
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Pose3d;
-import io.github.roboblazers7617.classylimelights.targets.RawFiducialTarget;
+import io.github.roboblazers7617.limelight.targets.RawFiducialTarget;
 
 /**
  * Represents a 3D Pose Estimate.
